Use Requests Sessions
This commit combines f654ef9b and c4a07e23

Use the requests.Session() object to prevent reopening the connection to
the Snipe-IT or JPS services every time we make a request.

From the Requests documentation:
"if you’re making several requests to the same host, the underlying TCP
connection will be reused, which can result in a significant
performance increase" [1]

In jamf2snipe, the script takes 5 minutes less when these changes are
applied on my decent internet connection.

[1] https://requests.readthedocs.io/en/v2.9.1/user/advanced/