Use more appropriate comparison operators
Python warns about the use of the "is" operator in the contexts that
I've replaced here. The issue boils down to "is" comparing two objects'
identities for equality, while we are looking for two objects' values
for equality.

The problem with this can be illustrated with the following example:

>>> str1 = "wow"
>>> str2 = "wow"
>>> str3 = "".join(["w", "o", "w"])
>>> str1 is str2
True
>>> str1 is str3
False

The "is" operator only worked in these situations by chance due to
optimizations in the Python interpreter.

https://stackoverflow.com/questions/2988017/string-comparison-in-python-is-vs