Change out dangerous default argument
Using '[]' as a default argument is dangerous since default arguments
are parsed at startup, not at function runtime. This causes 'previous'
to be set to a single list object at startup which will never change.
Objects added to or removed from the list will persist on any run where
no argument to 'previous' is given.

In this case, if 'get_snipe_users' was run more than once without
arguments, it would remember its list of users from the previous runs.