Create a request handler for jamf requests too
The request handler for Snipe-IT reduces code copying for handling of
rate limits. This new request handler does it for jamf requests. This
reduces duplication.

By adding the request handler to the session instead of every request,
we also reduce the arguments to each request.

Since the setup of each session object is now a bit more difficult, add
it to a function and make each instance of session creation use those
functions.