Put most non-global code in main() function
If something doesn't need to set variables needed by the functions, it
doesn't need to be in the global scope at all.

This reduces warnings about redefining variable names and brings us
closer to being able to import jamf2snipe from another module.

Now there's no need for SNIPE_USERS to be uppercase and CustomField
is flagged as an invalid name, so fix it too.